refactor(billsplit): extract shared receipt model to feature_billsplit_core (#1111 slice 4)
Airo Coins' expense screens reached directly into another app feature's
internals (bill_split/domain/models/receipt_item.dart) — one of the
cross-feature couplings blocking the coins extraction. Move the model
into a small feature_billsplit_core package that both features consume;
the old path keeps a re-export shim.

The model was already pure (equatable + the currency formatter, itself
now in core_app_shell), so no logic changed.

Remaining coins→bill_split coupling, deferred to slice 5: the expense
screens still import receipt_parser_service and itemized_split_screen
directly. Those carry heavy app-bound deps (Gemini Nano, ML Kit,
path_provider) and belong behind injected callbacks when the coins
presentation layer moves, not in a shared package.

docs-not-needed: extraction with a compatibility shim, no behavior change.
