+++ app/pubspec.yaml
@@ -90,6 +90,8 @@ dependencies:
     path: ../packages/airo
   feature_coin:
     path: ../packages/feature_coin
+  feature_billsplit_core:
+    path: ../packages/feature_billsplit_core
   feature_coins_core:
     path: ../packages/feature_coins_core
