+++ packages/feature_billsplit_core/pubspec.yaml
@@ -1,0 +1,24 @@
+name: feature_billsplit_core
+description: >
+  Shared bill-split domain models (receipt items and their line-level split
+  ownership) consumed by both the Bill Split feature and Airo Coins.
+version: 0.0.1
+publish_to: none
+
+environment:
+  sdk: ^3.12.2
+  flutter: ">=1.17.0"
+
+dependencies:
+  core_app_shell:
+    path: ../core_app_shell
+  equatable: ^2.0.8
+  flutter:
+    sdk: flutter
+
+dev_dependencies:
+  flutter_test:
+    sdk: flutter
+  flutter_lints: ^6.0.0
+
+flutter:
